Recognize country that was clicked.
For now, just get the closest country to the click event.
Also, CORS in Chrome makes it real difficult to read a regular old JSON file in the same directory when testing in localhost (which is where I test, duh). So I'm using an old trick, and changing my JSON file to a JS file that sets the variable that I would have dumped the JSON data to.